Move cubit wrapper to submodule
