+++ cubitpy/cubit_wrapper/__init__.py
@@ -28,28 +28,5 @@
 # SOFTWARE.
 # -----------------------------------------------------------------------------
 """
-This module contains extensions and utility functions for the cubit python
-interface.
+This module defines a wrapper for cubit python interpreters.
 """
-
-# Python modules.
-import sys
-
-# Global options for cubitpy.
-from .conf import cupy
-
-# Global object.
-from .cubitpy import CubitPy
-
-# Utility functions.
-from .cubit_utility import get_surface_center
-
-# Define the itCouplingems that will be exported by default.
-__all__ = [
-    # cubit options.
-    "cupy",
-    # Cubit objects.
-    "CubitPy"
-    # Utility functions.
-    "get_surface_center",
-]
